Create OMEROSession interface and implement
The previous OMEROSession becomes DefaultOMEROSession.
